Fix import error, cleanup files
