+++ react-fe/package.json
@@ -8,8 +8,11 @@
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
+    "firebase": "^8.6.7",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-firebaseui": "^5.0.2",
+    "react-router": "^5.2.0",
     "react-scripts": "4.0.3",
     "web-vitals": "^1.0.1",
     "yarn": "^1.22.10"
