+++ react-fe/src/components/Constants.js
@@ -1,0 +1,5 @@
+export const Constants = {
+  imgAlt: {
+    userParking: "A user-uploaded image of the parking spot for rent"
+  }
+};
