--- react-fe/src/components/Home.js
@@ -1,12 +1,0 @@
-import React from 'react'
-import './Home.css'
-import Banner from './Banner'
-
-function Home() {
-    return (
-        <div className='home'>
-            <Banner />
-        </div>
-    )
-}
-export default Home
