+++ react-fe/src/components/organisms/Header.jsx
@@ -1,9 +1,9 @@
-import React from 'react'
-import './Header.css'
-import SearchIcon from "@material-ui/icons/Search"
-import ExpandMoreIcon from "@material-ui/icons/ExpandMore"
-import LanguageIcon from "@material-ui/icons/Language"
-import { Avatar} from "@material-ui/core";
+import React from 'react';
+import '../../styling/Header.css';
+import SearchIcon from "@material-ui/icons/Search";
+import ExpandMoreIcon from "@material-ui/icons/ExpandMore";
+import LanguageIcon from "@material-ui/icons/Language";
+import { Avatar } from "@material-ui/core";
 
 
 function Header() {
@@ -28,7 +28,7 @@ function Header() {
                 <Avatar />
             </div>
         </div>
-    )
+    );
 }
 
-export default Header
+export default Header;
