+++ react-fe/src/components/organisms/Navbar.jsx
@@ -1,7 +1,7 @@
 import { makeStyles } from '@material-ui/core/styles';
 import { Button, AppBar, Toolbar, IconButton, Typography } from '@material-ui/core';
 import MenuIcon from '@material-ui/icons/Menu';
-import RegisterForm from '../molecules/RegisterForm';
+import RegisterForm from './RegisterForm';
 
 const Navbar = () => {
   const useStyles = makeStyles((theme) => ({
