+++ react-fe/src/components/pages/Home.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import '../../styling/Home.css';
+import Banner from '../organisms/Banner';
+
+function Home() {
+    return (
+        <div className='home'>
+            <Banner />
+        </div>
+    );
+}
+export default Home;
