+++ react-fe/src/index.js
@@ -1,19 +1,20 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
+import './styling/index.css';
 import App from './App';
+import { Switch, Route } from 'react-router';
 import reportWebVitals from './reportWebVitals';
-import Login from './components/login/Login'
+import Login from './components/organisms/login/Login';
 
 ReactDOM.render(
-  <React.StrictMode>
-      <App>
-          <Switch>
-              <Route exact path="/login" component={Login}></Route>
-          </Switch>
-      </App>
-  </React.StrictMode>,
-  document.getElementById('root')
+    <React.StrictMode>
+        <App>
+            <Switch>
+                <Route exact path="/login" component={Login}></Route>
+            </Switch>
+        </App>
+    </React.StrictMode>,
+    document.getElementById('root')
 );
 
 // If you want to start measuring performance in your app, pass a function
