+++ react-fe/src/styling/Search.css
